Move adjust methods to layman.common.micka.util
